Changed trends + update data
Changed the trends.rmd so we have the correct years in the data. I also kept the original coding of the

One problem with using geom.smooth() is that is calculates a score for the year 2015, which we actually do not have in the data. It still calculates these, which looks a bit ugly.